so close but not quite .. string interp with json is annoying
